doc: add interceptors, identify user from middle and autorize operations

+++ src/home/home.controller.ts
@@ -1,8 +1,12 @@
-import { PropertyType } from '@prisma/client';
+import { PropertyType, UserType } from '@prisma/client';
 import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dto/home.dto';
 import { HomeService } from './home.service';
-import { Body, Controller, Delete, Get, Post, Put, Query, Param, ParseIntPipe } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Post, Put, Query, Param, ParseIntPipe, UnauthorizedException } from '@nestjs/common';
 import { User } from 'src/user/decorators/user.decorator';
+import { UserInterceptorType } from 'src/user/interceptors/user.interceptor';
+
+
+
 
 @Controller('home')
 export class HomeController {
@@ -33,26 +37,42 @@ export class HomeController {
   }
 
   @Get(':id')
-  getHome() {
-    return {};
+  getHome(@Param('id', ParseIntPipe) id: number) {
+    return this.homeService.getHomeById(id);
   }
 
+  // @Roles(UserType.REALTOR, UserType.ADMIN)
   @Post()
-  createHome(@Body() body: CreateHomeDto, @User() user) {
-    return console.log(user)
-    return this.homeService.createHome(body);
+  createHome(@Body() body: CreateHomeDto, @User() user: UserInterceptorType) {
+    return this.homeService.createHome(body, user.id);
   }
 
   @Put(':id')
-  updateHome(
+  async updateHome(
     @Param('id', ParseIntPipe) id: number,
     @Body() body: UpdateHomeDto,
+    @User() user: UserInterceptorType,
   ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    }
+
     return this.homeService.updateHomeById(id, body);
   }
 
   @Delete(':id')
-  deleteHome(@Param('id', ParseIntPipe) id: number) {
+  async deleteHome(
+    @Param('id', ParseIntPipe) id: number,
+    @User() user: UserInterceptorType,
+  ) {
+    const realtor = await this.homeService.getRealtorByHomeId(id);
+
+    if (realtor.id !== user.id) {
+      throw new UnauthorizedException();
+    }
+
     return this.homeService.deleteHomeById(id);
   }
 }
